Fix "line mode" diffs.
In a "line mode" diff, the code hashed each input line to a unique
character before performing the diff. However, the range of the hash
output was restricted to the character set of the input strings (256
values for char; 65,536 for 16-bit wchar_t), causing a diff failure when
the input size exceeded that many lines.

diff_lineMode now hashes each input line to a size_t value, which is
sufficient to accommodate any string that can fit in memory.